requirements.txt
streamlit
openai
langchain
python-dotenv
tiktoken